pyproject.toml
[project]
name = "jobs-tracker"
version = "2.0.0"
description = "Multi-employer Swiss job tracker — daily automated career monitoring for Novartis, Sandoz, Roche, BIS, and ECB"
readme = "README.md"
requires-python = ">=3.10.13"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "feedparser>=6.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "rich>=13.0",
    "python-dateutil>=2.9",
    "brotli>=1.1",
]

[dependency-groups]
dev = ["pytest>=8.0"]

[project.scripts]
jobs-tracker = "jobs_tracker.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jobs_tracker"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
